+++ src/lib/Bootstrap/Interactivity/Molecule.ts
@@ -171,6 +171,7 @@ namespace LiteMol.Bootstrap.Interactivity.Molecule {
     
     export function isMoleculeModelInteractivity(info: Info): info is Info.Selection {
         if (Interactivity.isEmpty(info)) return false;
+        if (info.source.type.info.typeClass === Entity.VisualClass && info.source.type !== Entity.Molecule.Visual) return false;
         let modelOrSelection = Utils.Molecule.findModelOrSelection(info.source);
         if (!modelOrSelection) return false;
         return true;