+++ src/lib/Visualization/CHANGELOG.md
@@ -1,6 +1,7 @@
 ﻿# 1.7.0
 * Added support for labels.
 * Added support for additional primitives.
+* Updated Surface primitive transforms.
 
 # 1.6.7
 * Ability to customize maximum H bond length.